requirements.txt
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
configparser==3.5.0
cryptography==2.1.3
cycler==0.10.0
fhirclient==3.2.0
idna==2.6
isodate==0.6.0
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
opencv-python==3.3.0.10
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.4.0
pyparsing==2.2.0
python-aiml==0.9.1
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
